% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{default_i10_iss_method}
\alias{default_i10_iss_method}
\title{The default ICD 10 ISS method}
\usage{
default_i10_iss_method()
}
\value{
The default ICD 10 ISS method to be used as a character string.
}
\description{
This function looks up the i10_iss_method environment variable which may be set by adding `i10_iss_method=roc_max_NIS` to
the .Rprofile file. The function also informs the user of the difference between the methods so an informed choice can be made.
}
