% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{injury}
\alias{injury}
\title{ICD10CM injury codes}
\format{
A data frame with 100,477 rows and 11 variables:
\describe{
  \item{dx1}{1st ICD-10-CM injury code recorded on an encounter.}
  \item{dx2}{2nd ICD-10-CM injury code recorded on an encounter.}
  \item{dx3}{3rd ICD-10-CM injury code recorded on an encounter.}
  \item{dx4}{4th ICD-10-CM injury code recorded on an encounter.}
  \item{dx5}{5th ICD-10-CM injury code recorded on an encounter.}
  \item{dx6}{6th ICD-10-CM injury code recorded on an encounter.}
  \item{dx7}{7th ICD-10-CM injury code recorded on an encounter.}
  \item{dx8}{8th ICD-10-CM injury code recorded on an encounter.}
  \item{dx9}{9th ICD-10-CM injury code recorded on an encounter.}
  \item{dx10}{10th ICD-10-CM injury code recorded on an encounter.}
  \item{died}{A binary indicator variable for death. 1 = died. 0 = survived.}
  ...
}
}
\usage{
injury
}
\description{
A dataset containing a sample of trauma registry for use in examples and tests.
}
\keyword{datasets}
