% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma.R
\name{trauma}
\alias{trauma}
\title{Trauma function}
\usage{
trauma(df, dx_pre, calc_method = 1, icd10 = T,
  conflict_resolution = "max")
}
\arguments{
\item{df}{dataframe containing ICD9 codes with common prefix}

\item{dx_pre}{prefix for diagnosis codes (example: dx1, dx2, ect)}

\item{calc_method}{ISS calculation method}
}
\description{
For each observation, program to assign severity and ISS body region values to each valid ICD-9-CM trauma code, assign Barell and AP component categories to each valid ICD-9-CM trauma code, calculate injury severity score (ISS) and new injury severity score (NISS), assign major mechanism, minor mechanism and intent for up to 4 E-Codes (excluding E-Code place) and assign trauma type (blunt or penetrating) based on major mechanism  of the first E-Code found.
}

